// jobup-frontend/src/app/generated-sources/openapi/model/workerResponseDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO returned in responses to the frontend
 */
export interface WorkerResponseDto { 
    id?: string;
    fullName?: string;
    jobType?: string;
    phoneNumber?: string;
    location?: string;
    rating?: number;
    description?: string;
}